backend/pyproject.toml
[project]
name = "model-test-backend"
version = "0.1.0"
description = "Model testing backend for FarmOS shopping mall CS chatbot"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.9",
    "pdfplumber>=0.11.0",
    "pymupdf>=1.24.0",
    "easyocr>=1.7.0",
    "pillow>=10.0.0",
    "python-docx>=1.1.0",
    "anthropic>=0.49.0",
    "openai>=1.58.0",
    "chromadb>=0.5.0",
    "python-dotenv>=1.0.1",
    "sse-starlette>=2.1.0",
    "httpx>=0.27.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.0.0",
    "ollama>=0.6.1",
    "google-genai>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
